+++ .eslintrc.js
@@ -51,6 +51,9 @@ module.exports = {
         '@typescript-eslint/no-unused-vars': 'error',
         '@typescript-eslint/explicit-function-return-type': 'error',
         '@typescript-eslint/no-explicit-any': 'error',
+
+        // We use noop functions in many places, for example as default values for props.
+        '@typescript-eslint/no-empty-function': 'off',
     },
     overrides: [
         {