+++ packages/thumbprint-codemods/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+## 0.5.1 - 2020-06-18
+
+### Changed
+
+-   [Patch] Upgrade version of Prettier.
+
 ## 0.5.0 - 2020-01-09
 
 ### Added