+++ packages/thumbprint-font-face/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+## 1.0.15 - 2020-07-14
+
+### Fixed
+
+-   [Patch] Upgrade Thumbprint Tokens to latest version. The changes do not affect this package.
+
 ## 1.0.14 - 2020-01-09
 
 ### Changed