+++ packages/thumbprint-global-css/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+## 0.1.17 - 2020-07-14
+
+### Fixed
+
+-   [Patch] Upgrade Thumbprint Tokens to latest version. The changes do not affect this package.
+
 ## 0.1.16 - 2019-10-17
 
 ### Changed