+++ packages/thumbprint-scss/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+## 4.0.1 - 2020-07-14
+
+### Fixed
+
+-   [Patch] Upgrade Thumbprint Tokens to latest version. This pulls in small color tweaks to improve accessibility.
+
 ## 4.0.0 - 2019-11-21
 
 ### Removed